refactor(setup): move YAML-specific logic to yaml.go
